requirements.txt
numpy
sounddevice
websockets
python-dotenv
python-evdevwebrtcvad